+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Unless otherwise specified, any version comparison below is the comparison of th
   - `PressurePlateActivateSound` and `PressurePlateDeactivateSound` -> `PressurePlateSound(blockState, activated)`
   - `DoorOpenSound` and `DoorCloseSound` -> `DoorSound(blockState, open)`
   - `TrapdoorOpenSound` and `TrapdoorCloseSound` -> `TrapdoorSound(blockState, open)`
+  - `FenceGateOpenSound` and `FenceGateCloseSound` -> `FenceGateSound(blockState, open)`
 - Merged world tick thread and network thread into a single thread using an event-driven wake-up mechanism. This simplifies
   the threading model while maintaining low packet processing latency through `LockSupport.parkNanos/unpark`.